Server/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bert-score==0.3.13
bidict==0.23.1
blinker==1.7.0
bokeh==3.4.1
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
chainlit==0.7.0
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
dataclasses-json-speakeasy==0.5.11
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docarray==0.32.1
docopt==0.6.2
emoji==2.11.1
entrypoints==0.4
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.99.1
fastapi-socketio==0.0.10
filelock==3.13.4
filetype==1.2.0
Flask==3.0.3
Flask-Testing==0.8.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
giskard==2.12.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
griffe==0.45.1
grpcio==1.62.2
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.49
Lazify==0.4.0
libmagic==1.0
literalai==0.0.504
llvmlite==0.42.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mixpanel==4.10.1
mkl==2021.4.0
mlflow-skinny==2.13.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.23.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
prisma==0.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.15
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.12
pyparsing==3.1.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.6
python-socketio==5.11.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.7
rapidfuzz==3.8.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
sentry-sdk==2.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.33.0
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.41.0
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
unstructured==0.13.3
unstructured-client==0.18.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.23.2
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
